requirements.txt
ipykernel
ipywidgets
numpy
pandas
pyarrow
python-dotenv
boto3
openai
langchain
langchain-openai
langchain-google-vertexai
langgraph
langsmith
langchain-aws
langchain-community
aws-requests-auth
pymupdf
pillow
frontend
faiss-cpu
python-docx
langchain-unstructured
unstructured[pdf]
layoutparser
streamlit
PyPDF2
vertexai
presidio-analyzer
presidio-anonymizer
pytesseract
requests
opencv-python-headless